Form Class: Separated Input class from Element inheritance
