+++ src/Forms/Input/Checkbox.php
@@ -19,7 +19,6 @@ along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 namespace Gibbon\Forms\Input;
 
-use Gibbon\Forms\Element;
 use Gibbon\Forms\Traits\MultipleOptionsTrait;
 
 /**
@@ -28,7 +27,7 @@ use Gibbon\Forms\Traits\MultipleOptionsTrait;
  * @version v14
  * @since   v14
  */
-class Checkbox extends Element
+class Checkbox extends Input
 {
     use MultipleOptionsTrait;
 
@@ -61,7 +60,7 @@ class Checkbox extends Element
         return ($this->value[$index] == 1 || $this->value[$index] == '1' || strtolower($this->value[$index]) == 'true' )? 'checked' : '';
     }
 
-    protected function getElement()
+    public function getOutput()
     {
         $output = '';
